--- a/conjure_oxide/Cargo.toml
+++ b/conjure_oxide/Cargo.toml
@@ -13,7 +13,7 @@ conjure_core = { path = "../crates/conjure_core" }
 minion_rs = { path = "../solvers/minion" }
 
 
-uniplate = "0.1.5"
+uniplate = "0.2.0"
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.137"
 serde_with = "3.12.0"