--- a/crates/conjure_core/Cargo.toml
+++ b/crates/conjure_core/Cargo.toml
@@ -11,7 +11,7 @@ conjure_macros = { path = "../conjure_macros" }
 enum_compatability_macro = { path = "../enum_compatability_macro" }
 minion_rs = { path = "../../solvers/minion" }
 
-uniplate = "0.1.5"
+uniplate = "0.2.0"
 project-root = "0.2.2"
 linkme = "0.3.31"
 serde = { version = "1.0.217", features = ["derive"] }