--- a/crates/conjure_core/src/ast/symbol_table.rs
+++ b/crates/conjure_core/src/ast/symbol_table.rs
@@ -4,7 +4,6 @@ use std::fmt::Display;
 use serde::{Deserialize, Serialize};
 
 use crate::ast::variables::DecisionVariable;
-use uniplate::{Biplate, Uniplate};
 
 #[derive(Clone, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum Name {