- **update uniplate to v0.2.0**
  PERFORMANCE
  
  This Uniplate update brought optimisations to `contexts` and
  `contexts_bi`.
  
  Using the methodology given in c4c4d7d3a (perf: only log rule not
  applicable in debug mode, 2025-01-16), and the `optimisations
  /implies-tautologies-cse` test case, the impact of this change is the
  following:
  
  Main (075c952): 790.9ms +- 11.5ms
  This PR:        730.2ms +- 18.7ms
  Savile Row -O0: 139.9ms +- 12.0ms
  
  (wall time)
  
  This is a 7.68% wall-time improvement compared to main.
  

- **tidy: use simpler `Biplate` syntax where possible**
  As discussed in https://github.com/conjure-cp/uniplate/issues/19.