@ozgurakgun 2025-01-30
I wonder how much of the time is spent on parsing for Oxide. Is it the via-conjure parser or the tree-sitter one you tested with?